+++ src/main/java/com/chriscarini/jetbrains/locchangecountdetector/LoCService.java
@@ -4,6 +4,7 @@ import com.chriscarini.jetbrains.locchangecountdetector.data.ChangeInfo;
 import com.chriscarini.jetbrains.locchangecountdetector.factory.LOCBaseWidgetFactory;
 import com.chriscarini.jetbrains.locchangecountdetector.git.GitNumStat;
 import com.chriscarini.jetbrains.locchangecountdetector.messages.Messages;
+import com.intellij.idea.ActionsBundle;
 import com.intellij.notification.Notification;
 import com.intellij.notification.NotificationGroupManager;
 import com.intellij.notification.NotificationType;
@@ -19,13 +20,19 @@ import com.intellij.openapi.project.Project;
 import com.intellij.openapi.roots.ProjectRootManager;
 import com.intellij.openapi.util.Disposer;
 import com.intellij.openapi.util.Pair;
-import com.intellij.openapi.vcs.actions.commit.CommonCheckinProjectAction;
+import com.intellij.openapi.vcs.FilePath;
+import com.intellij.openapi.vcs.ProjectLevelVcsManager;
+import com.intellij.openapi.vcs.actions.commit.CheckinActionUtil;
+import com.intellij.openapi.vcs.changes.LocalChangeList;
 import com.intellij.openapi.vfs.VirtualFileManager;
 import com.intellij.openapi.vfs.newvfs.BulkFileListener;
 import com.intellij.openapi.vfs.newvfs.events.VFileEvent;
 import com.intellij.util.ui.update.MergingUpdateQueue;
 import com.intellij.util.ui.update.Update;
+import com.intellij.vcsUtil.VcsUtil;
 import git4idea.commands.GitCommand;
+import java.util.Arrays;
+import java.util.stream.Collectors;
 import org.jetbrains.annotations.NonNls;
 import org.jetbrains.annotations.NotNull;
 
@@ -171,20 +178,21 @@ public class LoCService implements Disposable {
 
         @Override
         public void actionPerformed(@NotNull AnActionEvent e) {
-            // Pull up the commit dialog...
-            final CommonCheckinProjectAction f = new CommonCheckinProjectAction();
-            f.actionPerformed(e);
-//            NOTE: THE BELOW IS A COPY OF THE ABOVE TO AVOID `./gradlew verifyPlugin` FROM FAILING FOR "1 override-only API usage violation". 
-//            WRITTEN & COMMENTED OUT ON 2024-11-14 AND SIMPLY JUST SETTING `OVERRIDE_ONLY_API_USAGES` FOR `verifyPlugin` TASK.  
-//            // The below was pulled from the underlying implementation of `CommonCheckinProjectAction.actionPerformed(e)` on 2024-11-14 to avoid
-//            // `./gradlew verifyPlugin` from failing for "1 override-only API usage violation".
-//            final List<FilePath> roots = Arrays.stream(ProjectLevelVcsManager.getInstance(myProject).getAllVcsRoots())
-//                .filter((it) -> it.getVcs() != null && it.getVcs().getCheckinEnvironment() != null)
-//                .map((it) -> getFilePath(it.getPath()))
-//                .collect(Collectors.toList());
-//            final LocalChangeList initialChangelist = CheckinActionUtil.INSTANCE.getInitiallySelectedChangeList(myProject, e);
-//  
-//            CheckinActionUtil.INSTANCE.performCommonCommitAction(e, myProject, initialChangelist, roots, ActionsBundle.message("action.CheckinProject.text"), null, false);
+//            // Pull up the commit dialog...
+//            final CommonCheckinProjectAction f = new CommonCheckinProjectAction();
+//            f.actionPerformed(e);
+//            NOTE: THE ABOVE IS A SHORTHAND OF THE BELOW, BUT ABOVE CAUSES `./gradlew verifyPlugin` TO FAIL FOR "1 override-only API usage violation". 
+//            ~WRITTEN & COMMENTED OUT ON 2024-11-14 AND SIMPLY JUST SETTING `OVERRIDE_ONLY_API_USAGES` FOR `verifyPlugin` TASK.~  
+//            ABOVE WAS COMMENTED OUT ON 2025-01-19 SINCE THE PLUGIN COULD NO LONGER PASS `verifyPlugin` (`OVERRIDE_ONLY_API_USAGES`) TASK.  
+            // The below was pulled from the underlying implementation of `CommonCheckinProjectAction.actionPerformed(e)` on 2024-11-14 to avoid
+            // `./gradlew verifyPlugin` from failing for "1 override-only API usage violation".
+            final List<FilePath> roots = Arrays.stream(ProjectLevelVcsManager.getInstance(myProject).getAllVcsRoots())
+                .filter((it) -> it.getVcs() != null && it.getVcs().getCheckinEnvironment() != null)
+                .map((it) -> VcsUtil.getFilePath(it.getPath()))
+                .collect(Collectors.toList());
+            final LocalChangeList initialChangelist = CheckinActionUtil.INSTANCE.getInitiallySelectedChangeList(myProject, e);
+
+            CheckinActionUtil.INSTANCE.performCommonCommitAction(e, myProject, initialChangelist, roots, ActionsBundle.message("action.CheckinProject.text"), null, false);
 
             final Consumer<ChangeInfo> callback = ChangeThresholdService.getInstance(myProject).getCreateCommitActionCallback();
             if (callback != null) {